+++ CHANGELOG.md
@@ -23,6 +23,8 @@ and this project adheres to
   - [#1881](https://github.com/iovisor/bpftrace/pull/1881)
 - Automatic resolution of library paths for uprobes
   - [#1971](https://github.com/iovisor/bpftrace/pull/1971)
+- Support positional parameters as integer literals
+  - [#1982](https://github.com/iovisor/bpftrace/pull/1982)
 
 #### Changed
 - Prevent LLVM from unrolling loops